restructure scripts
